scripts/requirements.txt
httpx>=0.24.0
motor>=3.1.1
papaparse>=5.3.0
psycopg2-binary>=2.9.5
PyJWT>=2.6.0
pytest>=7.3.1
pytest-asyncio>=0.21.0
python-dotenv>=0.21.0
redis>=4.5.4
dnspython>=2.4.0